refactor imports and tests
