+++ sktime/dists_kernels/distances/compose_tab_to_panel.py
@@ -123,7 +123,7 @@ class AggrDist(BasePairwiseTransformerPanel):
     def get_test_params(cls, parameter_set="default"):
         """Test parameters for AggrDist."""
         # importing inside to avoid circular dependencies
-        from sktime.dists_kernels import ScipyDist
+        from sktime.dists_kernels.distances.scipy_dist import ScipyDist
 
         return [
             {"transformer": ScipyDist(), "aggfunc_is_symm": True},
@@ -204,6 +204,6 @@ class FlatDist(BasePairwiseTransformerPanel):
     def get_test_params(cls, parameter_set="default"):
         """Test parameters for FlatDist."""
         # importing inside to avoid circular dependencies
-        from sktime.dists_kernels import ScipyDist
+        from sktime.dists_kernels.distances.scipy_dist import ScipyDist
 
         return {"transformer": ScipyDist()}
