+++ sktime/dists_kernels/distances/tests/__init__.py
@@ -1,0 +1,2 @@
+# -*- coding: utf-8 -*-
+"""Distances tests."""
