+++ sktime/dists_kernels/distances/tests/test_scipy_dist.py
@@ -1,8 +1,9 @@
 # -*- coding: utf-8 -*-
+"""Test ScipyDist."""
 import numpy as np
 
+from sktime.dists_kernels.distances.scipy_dist import ScipyDist
 from sktime.utils._testing.panel import make_transformer_problem
-from sktime.dists_kernels.scipy_dist import ScipyDist
 
 X1 = make_transformer_problem(
     n_instances=5,
@@ -69,6 +70,7 @@ COLALIGN_VALUES = ["intersect", "force-align", "none"]
 
 
 def test_scipydist():
+    """Test ScipyDist."""
     # test numpy
     _run_scipy_dist_test(X1, X2)
